Made the tank turn function smarter
The tank turn function should now turn left if the number is greater then 180. For example, if the number give for angle is 90 it would make the robot turn 90 degrees to the right, and if its 270, it would turn 90 degrees to the left.

I also created a couple of functions to make it easier to write, I created two functions to shorten the motor_pair.move_tank_for_degrees and the motor_pair.move_tank commands to shorten what you have to type.